[Minor] Adding retry to gradient test cases to avoid failure
